unlink e altri elelmenti

+++ app/Http/Controllers/element_edit_controller.php
@@ -38,7 +38,7 @@ class element_edit_controller extends Controller
                 if(isset($request->file_name)){
                     //cancella foto
                     $path =  $element->pathPhoto;
-                    if(file_exists('storage'.$path)){
+                    if(file_exists('storage'.$path) && $path != null){
                         unlink(public_path('storage'.$path));
                     }
